requirements-django.txt
-r requirements-core.txt
django>=6.0
djangorestframework>=3.16
django-cors-headers>=4.9
drf-spectacular>=0.28
django-cryptography>=1.1
google-auth>=2.49
google-auth-oauthlib>=1.2
